Move evaluatable definitions to the plugin
