+++ src/PGroonga.EntityFrameworkCore/Query/ExpressionTranslators/Internal/PGroongaEvaluatableExpressionFilterPlugin.cs
@@ -1,0 +1,15 @@
+using System.Linq.Expressions;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Query;
+
+namespace PGroonga.EntityFrameworkCore.Query.ExpressionTranslators.Internal
+{
+    public class PGroongaEvaluatableExpressionFilterPlugin : IEvaluatableExpressionFilterPlugin
+    {
+        public bool IsEvaluatableExpression(Expression expression)
+            => !(
+                expression is MethodCallExpression exp
+                && exp.Method.DeclaringType == typeof(PGroongaDbFunctionsExtensions)
+            );
+    }
+}
