+++ tests/PGroongaTests/PGroongaTests.csproj
@@ -10,15 +10,15 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational.Specification.Tests" Version="3.1.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.0" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational.Specification.Tests" Version="5.0.0-*" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.0-*" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
-        <PackageReference Include="coverlet.msbuild" Version="2.8.0" />
+        <PackageReference Include="coverlet.msbuild" Version="2.9.0" />
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\PGroonga.EntityFrameworkCore\PGroonga.EntityFrameworkCore.csproj" />
+        <ProjectReference Include="..\..\src\PGroonga.EntityFrameworkCore\PGroonga.EntityFrameworkCore.csproj" />
     </ItemGroup>
 
 </Project>
